Rename contributes to contributions.
Most (all) keys are nouns, `contributes` is the only verb – I think, i'm
not a linguist.

I think for consistency it should be a noun as well.

I know that vscode uses contributes, so we can also decide to keep
contributes.

It may also be that the contributes feels weird to me as a non-native
speaker and that it's ok for a native one ? SO happy to not go this
route.