Simplify reader manipulation.
Migrate from a dict of list with complex mutations, with a flatter table
that we completely filter everytime.

Even with a few thousand readers this should still be quite fast and
simplify the code a bit. (Boring is good). We also make the reader
hashable even if technically the filename patterns are mutable (maybe
they should not be?) to simplify manipulation a bit more.