Simplify Reader/writer internal logic. (#107)
* Simplify reader manipulation.

Migrate from a dict of list with complex mutations, with a flatter table
that we completely filter everytime.

Even with a few thousand readers this should still be quite fast and
simplify the code a bit. (Boring is good). We also make the reader
hashable even if technically the filename patterns are mutable (maybe
they should not be?) to simplify manipulation a bit more.

* Simplify writers: remove interval tree.

Trying to manipulate interval-trees is IMHO more complex that what we
need to, and instead of having a dict-of-interval tree, it is almost as
simple to store a table and filter the table on removal.

It is a bit less efficient optimally in some cases, but that should
still be sufficiently fast even with thousands of writers/layer
constraints.

It also remove a dependency, which in this case is I believe a plus.
We could still go 1/2 way and have writers be a dict of (min,max,
writers) but we also don't have _that many_ layer types so it's not
going to be much more performant.

In the end if you squint this really look like sql, as we select writers
based on the layer types,bounds and order-by based on some categories,
So yes this is not sql as we have live objects, and q sql-like library
would be overkill... for now.

* Simplify iteration a bit.

We just start with all the writers and trim down.
Really in the end we are reinventing a sort-of sql.

* fix linters

* Update _docs/render.py

Co-authored-by: Talley Lambert <talley.lambert@gmail.com>