DOC Update and clarify supported widget contributions (#310)
This PR:

* Clarifies that the only supported widget parent classes are
`QtWidgets.QWidget` or `magicgui.widgets.Widget`. The "such as"
suggested that other parent classes could be supported
* Removes documented support for wrapper functions - going forward we do
not want to support wrapper functions. (I.e. we do not want new plugins
to use wrappers). it is difficult to type (both the wrapper, which will
just be a normal function, and the return object, which could be many
things) and is too flexible and difficult maintain.
* I think we want to be backwards compatible but this is outside the
scope of this PR and should be discussed in:
https://github.com/napari/napari/issues/6174

Context - this was noted as part of the menu contributions work
(https://github.com/napari/napari/issues/6174)