TST Fix `test_conversion_missing` (#341)
This test is failing in CIs in #340.

I think it is because there is no warning raised (not sure why the
`UserWarning` is there). It used to pass because of this bug in pytest
https://github.com/pytest-dev/pytest/issues/9036, but this was fixed in
https://github.com/pytest-dev/pytest/pull/11129 and included in a recent
release, thus our CIs are failing.

Note that I have update the test to be more specific about the exception
raised (`PackageNotFoundError`) but it is not needed, subclass
exceptions are still accepted by pytest. Also added a message to be more
specific.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>