Instantiate `_command_menu_map` with all needed `defaultdicts` (#369)
This PR fixes a bug in `napari/napari` with loading shimmed npe1
plugins.

On the napari side, we never distinguish between shimmed npe1 plugins
and npe2 plugins. This means even though npe1 plugins can't declare menu
contributions, we need to be able to index into the `_command_menu_map`
for any commands shimmed plugins declare, and get an empty list back.

This PR moves the instantiation of the two extra levels of `defaultdict`
up to the declaration of the `_command_menu_map`, so that it can always
be fully indexed. I've also added an additional comment, and a specific
test for npe1 adapters.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>