Fix naming of npe1 converted contribution commands and add tests (#374)
Prior to this PR, declaring multiple implementations of the same
hookspec in an npe1 plugin would lead to a converted manifest with
multiple commands of the same ID e.g. `plugin_name.get_reader` multiple
times, or `plugin_name.write_labels` multiple times.

Even though this is technically "valid" per the schema, it should not
be, because this leads to errors in trying to register the commands in
napari. App-model (correctly, imo), does not allow you to register
multiple commands with the same ID.

We should update the schema validator to disallow multiple declarations
of the same command ID, but in the meantime, this PR disambiguates
converted commands for the same hook spec by using the function name
instead of the spec name.

---------

Co-authored-by: Draga Doncila <ddon0001@student.monash.edu>