Show menus and submenus in the contributions page (#390)
Ok, this is a replacement for https://github.com/napari/npe2/pull/387
Instead of whack a mole I tried to go from the start looking at the
schema.json (release in this repo)
```
"menus": {
          "title": "Menus",
          "hide_docs": true,
          "type": "object",
          "additionalProperties": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "$ref": "#/definitions/MenuCommand"
                },
                {
                  "$ref": "#/definitions/Submenu"
                }
              ]
            }
          }
        },
        "submenus": {
          "title": "Submenus",
          "hide_docs": true,
          "type": "array",
          "items": {
            "$ref": "#/definitions/SubmenuContribution"
          }
        },
```
If you compare to some of the others:
```
        "readers": {
          "title": "Readers",
          "type": "array",
          "items": {
            "$ref": "#/definitions/ReaderContribution"
          }
        },
        "writers": {
          "title": "Writers",
          "type": "array",
          "items": {
            "$ref": "#/definitions/WriterContribution"
          }
        },
        "widgets": {
          "title": "Widgets",
          "type": "array",
          "items": {
            "$ref": "#/definitions/WidgetContribution"
          }
        },
```
So then it becomes more clear how to modify the template. i did get a
bit of help from Claude.
Also, I was working locally and had some issues with the schema.json
coming from the release, so hide kept being set.
I tweaked render.py to use local schema, which got things working for
me.

Let's see if CI likes it!

---------

Co-authored-by: Carol Willing <carolcode@willingconsulting.com>
Co-authored-by: Draga Doncila Pop <17995243+DragaDoncila@users.noreply.github.com>